[CFE] Fix scopes for expression compilation
Before this CL, when doing expression compilation we would build scopes
like this:

[exp scope locals] -> [exp scope imports] -> [org scope]

Which means that if one looks up "Foo" in the (first) scope and
something with that name is available both via imports and via
the original scope one would find it via the import.
That's not what we want considering the expression compilation
library should pretend to be the original library.

Instead this CL copies the original scope into the expression
compilation local scope:

[exp scope locals + org scope] -> [exp scope imports]

Meaning that we would find the correct "Foo" if one was defined in
the original library and (another one by the same name is) available
via an import.

Fixes https://github.com/dart-lang/sdk/issues/48169

Change-Id: Ib9e089b465f7da1a4c5799a2165fd06c735ea7da
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/229904
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>