Fix instance member lookup fallback for static property lookup
When a static property lookup fails, the analyzer's error recovery
logic tries to fall back on looking up an instance member.  An
unfortunate consequence of this is that the looked-up member may have
a type that refers to unbound type parameters.  For example:

    class C<T> {
      List<T> t = [];
    }
    var x = C.t; // (1)

Although the line at (1) has an error, the analyzer continues to
analyze it assuming the reference to `t` is valid, producing a static
type for `x` of `List<T>`.  But this isn't a valid type because `T`
has no meaning outside the context of the definition of class C.

To avoid unbound type parameters leaking outside the class, we do an
"instantiate to bounds" operation at the time of the error recovery;
for this example that results in a type `List<dynamic>` instead of
`List<T>`.

Change-Id: I763bf32d16345b0a1a6f9f6893e78a138e85c8c2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/230460
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>