[cfe] Don't create concrete forwarding stubs for invalid implementations
Creation of concrete forwarding stubs didn't take into account that an
override might add additional optional parameters making the overridden
member an invalid implementation for the override. This lead to a crash
when trying to compute the super parameter type for the additional
parameters.

Closes https://github.com/dart-lang/sdk/issues/48242

Change-Id: Ie46859d8de328a31eacf7905bb429b0bba61447f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/230663
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>