[cfe] Handle named arguments anywhere in enhanced enums
The arguments synthesized for enhanced enums didn't ensure correct
setup of `ArgumentsImpl.argumentsOriginalOrder` needed for the
named-arguments-anywhere feature.

The CL also handles local initializers as part of constants, as needed
when combining the two features.

Included is a fix in adding synthesized enum arguments to redirecting
initializers. This was accidentally done on all classes (not just enum
classes) when enhanced enums was enabled.

Closes https://github.com/dart-lang/sdk/issues/48276

Change-Id: If911e57bb338a50f2cee1721b4e43ff343685933
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/231440
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>