[cfe] Perform precompilation
Initial support for performing precompilations based on the computation
of [NeededPrecompilations].

Currently this requires the caller to supply a `macroSerializer` that
can convert a [Component] into a [Uri] that can be passed to the macro
executor.

Also, the caller needs to supply a `macroTarget` used for the
precompilation. This is done to avoid (for now) a dependency of the VM
target inside the CFE itself.

Change-Id: I50cc25fe6632a25c0c1a8a1ac019e6d9e35ab4f9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/231443
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>