Rewrite ParserDiagnosticsTest to use textual dumps of ASTs.
We can reuse the existing "resolved" AST printer here.

One possible deficiency in the existing implementation is that it
sorts children by name, does not print them in their syntactic
positions, i.e. for `(var a)` we should print `keyword: var`,
and then `identifier: a`. The reason for this sorting is that
implementation adds properties from superclasses later, and there
is no correct positions for it.

Of course we have `childrenEntities`, which are sorted correctly,
but they lack names. Maybe if we added a new API with names, we
could use it here.

Ah, another issue is resolution information, e.g. `staticElement`,
`staticType` - these don't have the right syntactic position. We
could add them after syntactic entities, this is probably not worse
than what we have now.

Change-Id: I1354556acd69afdce7da712ce10d1a1fc724871b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/232687
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>