Test an odd interaction of implicit `.call` tearoffs with promotion.
Currently, we decide whether to insert an implicit `.call` tearoff on
an expression based on the type the expression is being assigned to,
whereas the spec says that we decide it based on the context type.

In most cases, these are the same type, however when assigning to a
promoted variable, the context type is the promoted type, whereas the
"type being assigned to" is the unpromoted type.

Previously, we had no language tests of this behavior; it just emerged
naturally from the way it was implemented in the CFE.  This test locks
down the behavior so that we won't change it by mistake.

We may, at some future time, decide to change it on purpose though.
Discussion about this behavior is ongoing on the langauge team.

Change-Id: I82f54c8875fd3b5829c5f75a676fdaf5eea64703
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/233650
Reviewed-by: Lasse Nielsen <lrn@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>