Consolidate calls to InferenceContext.getContext.
This change moves all the call sites for `Inference.getContext` to the
top of `ResolverVisitor.visit` methods.  Any other methods that need
access to the inference context receive it as a parameter.

This is part of a larger effort to elimiate the use of
InferenceContext.getContext and InferenceContext.setType entirely.

Change-Id: Ie5b9dd3adf827a5b17b0f5662c5a7ed9c8e98069
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/233863
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>