Migration: avoid introducing a cast on an implicit `.call` tearoff.
The migration engine resolves the ASTs for the code in question twice,
once with null safety features turned off, in order to build the
nullability graph, and then a second time with null safety features
turned off, in order to determine the precise type of every
subexpression and figure out where `!`s and casts need to be inserted.

The analyzer passes type inference contexts around by storing them on
AST nodes.  As a result, if a context type is associated with an AST
node the first time the code is resolved, and no context is stored the
second time it is resolved, the leftover contextual information gets
erroneously used by the second resolution pass.  Before this change,
this was happening for expressions that received `.call` tearoffs,
because during the second resolution pass, the expression was wrapped
inside an `ImplicitCallReference` node, and the logic for analyzing
such nodes didn't bother to clear the context when visiting the
subexpression.

This change fixes the problem by directly calling `analyzeExpression`
to visit the subexpression (allowing the correct context to be
explicitly provided).

Fixes #47848.

Bug: https://github.com/dart-lang/sdk/issues/47848
Change-Id: Ie22c2821a0e04c1fb685d2cad7f8675fa08ef986
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/234080
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>