Move code locations computing when supportsLocationLink into DefinitionHandler.
The main reason to do this is that we add navigation regions while
visiting AST, so we will not able to call the async method to parse
a file.

An accidental improvement is also that we will not parse files
multiple times when the same element is referenced multiple times.
Although if different elements of the same file are referenced, we
don't have any improvement here.

Change-Id: Ic4d914e89515af31f2dc3c4f842f0313cc89fac8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/234863
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Danny Tuppeny <danny@tuppeny.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>