Use NamedType.name.staticElement in favor of DartType.alias. (dart-lang/linter#3279)
In some circumstances, the analyzer doesn't reliably set
DartType.alias to point to the type alias that led to a type.  (This
is the ultimate cause of dart-lang/linter#3275).  I'm currently researching whether
it's possible to improve the analyzer behavior in this regard.  In the
meantime, we can fix dart-lang/linter#3275 by changing the linter to grab the type
alias element directly from NamedType.name.staticElement.

Fixes dart-lang/linter#3275.