[cfe] Perform inference on enum elements as on other fields
The type inference on enum elements was performed at the outline
expression building phase, which is later than the inference on the
rest of the members. It prevents some type checks from being performed
correctly, for example, a check for the type of the getter induced by
the element and a declared static setter. This CL moves the inference
on enum elements to the same phase as for the rest of the class and
enum members.

Part of https://github.com/dart-lang/sdk/issues/47453

Change-Id: Iefc5606e6db07df8cd2de101c47d115c2ff8454a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/236000
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>