[cfe] Enable alternative invalidation strategy in incremental macro test
This adds support for changes to macro classes with the alternative
invalidation strategy. Because macro classes can modify code in any
library where they are applied, changes to libraries with macro classes,
or to libraries that they depend upon, are handled as a signature
change that require full recompilation of the dependencies.

A more precise approach would be to invalidate only the libraries that
apply the changed macros. This data is not currently available during
incremental compilation.

Change-Id: I23347e332b95acfe232f4a04a023677780a64d60
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/235681
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>