[cfe] Pass references to augmentation libraries
This adds support for preserving reference in augmentations.
Augmentation libraries are now passed `referencesFrom` such that
its members case use it.

A special case is classes. Classes need the `referencesFrom` both
for themselves and for their members but while injected classes should
use the reference from `referencesFrom` for their [Class] node,
augmentation classes shouldn't. This is because [Class] node for the
origin is used in the output, whereas the augmentation [Member] node is
used for the members.

Change-Id: Iea73a5edd8f6cd4801e7e8229583eab69d89968e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/236500
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>