Don't check consistency if an exception occurs during analysis.
A consistency check is not necessary in the event of an exception,
because we aren't returning a value anyhow.  And it's guaranteed to
fail, because the AnalysisDriver clears the context when an exception
occurs (see AnalysisDriver._clearLibraryContextAfterException).  This
is harmful, because the InconsistentAnalysisException covers up the
original exception, making debugging much more difficult.

Change-Id: Iaab1bdf1a32c1e261cef97b8749a2c5057d9b11c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/236780
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>